Fix js error with reordering
Happens in Immutable but traced back to api call of ajax put
The issue is due to the todo in question being an Immutable Map
rather than a js obj.
This happens due to the initial todo sync.  Todo sync pulls in
lists of todos and converts it to Immutable immediately.
This fix lets TodoListWrap handle this case, or the case where
each todo (in the store) was converted to js objects.