Refactor todo reorder async in reducer
Also fix some of the logic around reordering
there was a disparity between what the client does and what the server
does.  Since acceted todos and possibly completed todos could be
invisible, if the client drags a todo even a single space, there could
be a hidden todo that it's jumping over, therefore when the server is
told "jump 3 spaces" that could be inaccurate.

Have fixed the issue by making completion and acceptance of todos move
them to the bottom of the list.

There remains one problem though. If there is a list of todos like this:

t1: pos: 1
t2: pos: 4

and you call t1.move_lower
it then looks like this:

t1: pos: 2
t2: pos: 4

which obviously doesn't work, but also it contradicts the documentation
and even the code in the gemfile.  I'll leave it for now because I don't
want to be blocked