Cache data state in todolistwrap to avoid reprocessing
when any action is fired, even ones that won't affect todo list,
the subscribe() will fire the listener which is handleStateChange within
todolistwrap.  This caused, at the very least, grabbing the data out
from the store, turning it into a json, and reordering it according to
position.  Now we'll cache a copy of the data from the last time the
state was changed, then check it right away on state changes.