Prepping to add household name to appbar
But more importantly, fixing config error
when i used _.merge(options, emptystore)
emptystore was taking precedence so things were being set to null that
shouldn't be.  Not even sure how the app still worked in that case

As for the actual story for this commit, not sure i want to add
household name to appbar.  Tried doing it and there just isnt' room for
that information.  Need to rethink if there is much value there
Probably just adding it to the page title makes more sense