Step 1 of implement clear completed todos
Determine whether to render this button by reading the user permission
on the household.  Problem was, we just set the user permission
arbitrarily, using the default of isAdmin = false.

Now, created a getPermission api call which calls the server upon load
of the app (right after the initial sync todos request).